Comprobaciones arregladas

+++ EI1027NatureAdventure1/src/database/daoAvaliableBook.java
@@ -50,7 +50,7 @@ public class daoAvaliableBook implements DaoInterface {
 	
 	@Override
 	public void addElement(Object element) {
-		if(!AvaliableForBooking.class.equals(element)) return;
+		if(!(element instanceof AvaliableForBooking)) return;
 		AvaliableForBooking a = (AvaliableForBooking) element;
 		String sql = "INSERT INTO avaliableforbook (idact, ssnumber, date, schedule, avaliableplaces) "
 				+ "VALUES(?, ?, ?, ?, ?);";
@@ -59,6 +59,7 @@ public class daoAvaliableBook implements DaoInterface {
 
 	@Override
 	public void deleteElement(Object element) {
+		if(!(element instanceof AvaliableForBooking)) return;
 		AvaliableForBooking a = (AvaliableForBooking) element;
 		String sql = "DELETE FROM avaliableforbook WHERE ssnumber = ? AND date = ? AND schedule = ?;";
 		dataSource.update(sql, a.getSsNumber(), a.getDate(), a.getSchedule());
@@ -66,7 +67,7 @@ public class daoAvaliableBook implements DaoInterface {
 
 	@Override
 	public void updateElement(Object element) {
-		if(!AvaliableForBooking.class.equals(element)) return;
+		if(!(element instanceof AvaliableForBooking)) return;
 		AvaliableForBooking a = (AvaliableForBooking) element;
 		String sql = "UPDATE avaliableforbook SET avaliableplaces = ? WHERE ssnumber = ?, date = ?, schedule = ?;";
 		dataSource.update(sql, a.getAvaliablePlaces(), a.getSsNumber(), a.getDate(), a.getSchedule());
